+++ defenses/__init__.py
@@ -6,6 +6,7 @@ from .input_sanitizer import InjectionPatternFilter, MarkdownHeaderSanitizer, Un
 from .llm_classifier import LLMChunkClassifier
 from .output_validator import ExfiltrationGuard, InjectionSuccessGuard, StructuredOutputGuard
 from .tool_call_validator import ToolCallValidator
+from .cross_agent_boundary_guard import CrossAgentBoundaryGuard
 from .prompt_hardening import LEVELS, get_prompt
 
 __all__ = [
@@ -25,4 +26,5 @@ __all__ = [
     "InjectionSuccessGuard",
     "StructuredOutputGuard",
     "ToolCallValidator",
+    "CrossAgentBoundaryGuard",
 ]